chore: split integration tests to dedicated files
